Started "Orders" application Created Cart logic

+++ products/models.py
@@ -5,7 +5,7 @@ from django.utils.text import slugify
 class Product(models.Model):
     name = models.CharField(max_length=100, verbose_name='Название товара')
     slug = models.SlugField(max_length=100, allow_unicode=True, blank=True, unique=True)
-    price = models.DecimalField(decimal_places=2, max_digits=10, default=0, verbose_name='Цена')
+    price = models.FloatField(default=0, verbose_name='Цена')
     description = models.TextField(blank=True, verbose_name='Описание товара')
     quantity = models.PositiveIntegerField(verbose_name='Количество товара')
     image = models.ImageField(null=True, blank=True)
